+++ HabboHotel/Friends/MessengerDataLoader.cs
@@ -3,6 +3,7 @@ using Plus.Database;
 using Plus.HabboHotel.GameClients;
 using Plus.HabboHotel.Users;
 using Plus.HabboHotel.Users.Messenger;
+using Plus.Utilities;
 
 namespace Plus.HabboHotel.Friends;
 
@@ -107,7 +108,7 @@ internal class MessengerDataLoader : IMessengerDataLoader
     public async Task<Dictionary<int, List<(string Message, int SecondsAgo)>>> GetAndDeleteOfflineMessages(int userId)
     {
         using var connection = _database.Connection();
-        var messages = (await connection.QueryAsync<(int, string, int)>("SELECT from_id, message, timestamp FROM messenger_offline_messages WHERE to_id = @userId", new { userId })).GroupBy(r => r.Item1).ToDictionary(r => r.Key, r => r.OrderBy(r => r.Item3).Select(g => (g.Item2, (int)(PlusEnvironment.Now() - g.Item3))).ToList());
+        var messages = (await connection.QueryAsync<(int, string, int)>("SELECT from_id, message, timestamp FROM messenger_offline_messages WHERE to_id = @userId", new { userId })).GroupBy(r => r.Item1).ToDictionary(r => r.Key, r => r.OrderBy(r => r.Item3).Select(g => (g.Item2, (int)(UnixTimestamp.GetNow() - g.Item3))).ToList());
         await connection.ExecuteAsync("DELETE FROM messenger_offline_messages WHERE to_id = @userId", new { userId });
         return messages;
     }