+++ HabboHotel/IGame.cs
@@ -1,60 +1,8 @@
-using Plus.HabboHotel.Achievements;
-using Plus.HabboHotel.Bots;
-using Plus.HabboHotel.Cache;
-using Plus.HabboHotel.Catalog;
-using Plus.HabboHotel.Catalog.Utilities;
-using Plus.HabboHotel.Badges;
-using Plus.HabboHotel.GameClients;
-using Plus.HabboHotel.Games;
-using Plus.HabboHotel.Groups;
-using Plus.HabboHotel.Items;
-using Plus.HabboHotel.Navigator;
-using Plus.HabboHotel.Quests;
-using Plus.HabboHotel.Rooms;
-using Plus.HabboHotel.Rooms.Chat;
-using Plus.HabboHotel.Subscriptions;
-
 namespace Plus.HabboHotel;
 
 public interface IGame
 {
     void StartGameLoop();
     void StopGameLoop();
-
-    [Obsolete("Use dependency injection instead.")] IBadgeManager BadgeManager { get; }
-    [Obsolete("Use dependency injection instead.")] IGameClientManager ClientManager { get; }
-
-    [Obsolete("Use dependency injection instead.")] ICatalogManager Catalog { get; }
-
-    [Obsolete("Use dependency injection instead.")] INavigatorManager Navigator { get; }
-
-    [Obsolete("Use dependency injection instead.")] IItemDataManager ItemManager { get; }
-
-    [Obsolete("Use dependency injection instead.")] IRoomManager RoomManager { get; }
-
-    [Obsolete("Use dependency injection instead.")] IAchievementManager AchievementManager { get; }
-
-    [Obsolete("Use dependency injection instead.")] ISubscriptionManager SubscriptionManager { get; }
-
-    [Obsolete("Use dependency injection instead.")] IQuestManager QuestManager { get; }
-    [Obsolete("Use dependency injection instead.")] IQuestService QuestService { get; }
-    [Obsolete("Use dependency injection instead.")] ICatalogService CatalogService { get; }
-    [Obsolete("Use dependency injection instead.")] IAchievementService AchievementService { get; }
-
-    [Obsolete("Use dependency injection instead.")] IGroupManager GroupManager { get; }
-
-    [Obsolete("Use dependency injection instead.")] IChatManager ChatManager { get; }
-
-    [Obsolete("Use dependency injection instead.")] IGameDataManager GameDataManager { get; }
-
-    [Obsolete("Use dependency injection instead.")] IRoomService RoomService { get; }
-    [Obsolete("Use dependency injection instead.")] IRoomFactory RoomFactory { get; }
-    [Obsolete("Use dependency injection instead.")] IRoomAppender RoomAppender { get; }
-    [Obsolete("Use dependency injection instead.")] IItemService ItemService { get; }
-    [Obsolete("Use dependency injection instead.")] IBotUtility BotUtility { get; }
-    [Obsolete("Use dependency injection instead.")] IPetUtility PetUtility { get; }
-    [Obsolete("Use dependency injection instead.")] IBotManager BotManager { get; }
-
-    [Obsolete("Use dependency injection instead.")] ICacheManager CacheManager { get; }
     Task Init();
 }